+++ rust/source/src/connector_source.rs
@@ -1,17 +1,23 @@
 use std::fmt::Debug;
+use std::marker::Send;
+use std::sync::Arc;
 
+use async_trait::async_trait;
 use lazy_static::__Deref;
-use risingwave_common::array::StreamChunk;
+use risingwave_pb::data::StreamChunk;
 use risingwave_common::error::ErrorCode::{InternalError, ProtocolError};
 use risingwave_common::error::{Result, RwError};
 use risingwave_connector::base::SourceReader;
+use risingwave_connector::state;
+use risingwave_storage::memory::MemoryStateStore;
 
-use crate::{build_columns, SourceColumnDesc, SourceParser};
+use crate::{build_columns, SourceColumnDesc, SourceParser, StreamSourceReader};
 
+#[derive(Send)]
 pub struct ConnectorSource {
     pub parser: Box<dyn SourceParser>,
     pub reader: Box<dyn SourceReader>,
-    pub column_descs: Box<Vec<SourceColumnDesc>>,
+    pub column_descs: Vec<SourceColumnDesc>,
 }
 
 impl Debug for ConnectorSource {
@@ -70,5 +76,22 @@ impl ConnectorSource {
                 ))
             }
         }
+    }
+}
+
+#[derive(Debug)]
+pub struct ConnectorStreamSource {
+    source_reader: ConnectorSource,
+    state_store: state::SourceStateHandler<MemoryStateStore>,
+}
+
+#[async_trait]
+impl StreamSourceReader for ConnectorStreamSource {
+    async fn open(&mut self) -> Result<()> {
+        Ok(())
+    }
+
+    async fn next(&mut self) -> Result<StreamChunk> {
+        self.source_reader.next()
     }
 }
